Split commands/activity.rs (4,117 lines) into domain-specific submodules
Convert activity.rs into an activity/ directory with 11 domain-specific
submodules: db, logging, tokens, metrics, github_events, costs, velocity,
patterns, recommendations, budget, and timeline. All 40 Tauri command
signatures are preserved via barrel re-exports in mod.rs.

Also updates biome.json schema version from 2.3.12 to 2.3.13.

Closes #1883
