+++ src-tauri/src/commands/activity/tokens.rs
@@ -1,0 +1,134 @@
+use super::db::open_activity_db;
+use serde::{Deserialize, Serialize};
+
+/// Token usage summary by role
+#[derive(Debug, Serialize, Deserialize)]
+pub struct TokenUsageSummary {
+    pub role: String,
+    pub total_prompt_tokens: i64,
+    pub total_completion_tokens: i64,
+    pub total_tokens: i64,
+    pub activity_count: i32,
+    pub avg_tokens_per_activity: f64,
+}
+
+/// Query token usage statistics grouped by role
+#[tauri::command]
+pub fn query_token_usage_by_role(
+    workspace_path: &str,
+    since: Option<String>,
+) -> Result<Vec<TokenUsageSummary>, String> {
+    let conn =
+        open_activity_db(workspace_path).map_err(|e| format!("Failed to open database: {e}"))?;
+
+    let query = if since.is_some() {
+        "SELECT
+            a.role,
+            SUM(t.prompt_tokens) as total_prompt,
+            SUM(t.completion_tokens) as total_completion,
+            SUM(t.total_tokens) as total_tokens,
+            COUNT(DISTINCT a.id) as activity_count,
+            CAST(SUM(t.total_tokens) AS REAL) / COUNT(DISTINCT a.id) as avg_tokens
+         FROM agent_activity a
+         JOIN token_usage t ON a.id = t.activity_id
+         WHERE a.timestamp > ?1
+         GROUP BY a.role
+         ORDER BY total_tokens DESC"
+    } else {
+        "SELECT
+            a.role,
+            SUM(t.prompt_tokens) as total_prompt,
+            SUM(t.completion_tokens) as total_completion,
+            SUM(t.total_tokens) as total_tokens,
+            COUNT(DISTINCT a.id) as activity_count,
+            CAST(SUM(t.total_tokens) AS REAL) / COUNT(DISTINCT a.id) as avg_tokens
+         FROM agent_activity a
+         JOIN token_usage t ON a.id = t.activity_id
+         GROUP BY a.role
+         ORDER BY total_tokens DESC"
+    };
+
+    let mut stmt = conn
+        .prepare(query)
+        .map_err(|e| format!("Failed to prepare query: {e}"))?;
+
+    let summaries = if let Some(since_time) = since {
+        stmt.query_map([since_time], |row| {
+            Ok(TokenUsageSummary {
+                role: row.get(0)?,
+                total_prompt_tokens: row.get(1)?,
+                total_completion_tokens: row.get(2)?,
+                total_tokens: row.get(3)?,
+                activity_count: row.get(4)?,
+                avg_tokens_per_activity: row.get(5)?,
+            })
+        })
+        .map_err(|e| format!("Failed to query token usage: {e}"))?
+        .collect::<Result<Vec<_>, _>>()
+        .map_err(|e| format!("Failed to collect summaries: {e}"))?
+    } else {
+        stmt.query_map([], |row| {
+            Ok(TokenUsageSummary {
+                role: row.get(0)?,
+                total_prompt_tokens: row.get(1)?,
+                total_completion_tokens: row.get(2)?,
+                total_tokens: row.get(3)?,
+                activity_count: row.get(4)?,
+                avg_tokens_per_activity: row.get(5)?,
+            })
+        })
+        .map_err(|e| format!("Failed to query token usage: {e}"))?
+        .collect::<Result<Vec<_>, _>>()
+        .map_err(|e| format!("Failed to collect summaries: {e}"))?
+    };
+
+    Ok(summaries)
+}
+
+/// Token usage over time (daily aggregation)
+#[derive(Debug, Serialize, Deserialize)]
+pub struct DailyTokenUsage {
+    pub date: String,
+    pub role: String,
+    pub total_tokens: i64,
+}
+
+/// Query token usage over time, grouped by date and role
+#[tauri::command]
+pub fn query_token_usage_timeline(
+    workspace_path: &str,
+    days: Option<i32>,
+) -> Result<Vec<DailyTokenUsage>, String> {
+    let conn =
+        open_activity_db(workspace_path).map_err(|e| format!("Failed to open database: {e}"))?;
+
+    let days_val = days.unwrap_or(30);
+
+    let mut stmt = conn
+        .prepare(
+            "SELECT
+                DATE(a.timestamp) as date,
+                a.role,
+                SUM(t.total_tokens) as daily_tokens
+             FROM agent_activity a
+             JOIN token_usage t ON a.id = t.activity_id
+             WHERE a.timestamp > datetime('now', '-' || ?1 || ' days')
+             GROUP BY date, a.role
+             ORDER BY date DESC, a.role",
+        )
+        .map_err(|e| format!("Failed to prepare query: {e}"))?;
+
+    let timeline = stmt
+        .query_map([days_val], |row| {
+            Ok(DailyTokenUsage {
+                date: row.get(0)?,
+                role: row.get(1)?,
+                total_tokens: row.get(2)?,
+            })
+        })
+        .map_err(|e| format!("Failed to query timeline: {e}"))?
+        .collect::<Result<Vec<_>, _>>()
+        .map_err(|e| format!("Failed to collect timeline: {e}"))?;
+
+    Ok(timeline)
+}
